requirements.txt
# ── Telegram ────────────────────────────────────────────────
pyrogram==2.0.106
TgCrypto==1.2.5

# ── MongoDB ──────────────────────────────────────────────────
motor==3.4.0
pymongo[srv]==4.7.3

# ── Web server (Render health endpoint) ──────────────────────
aiohttp==3.9.5

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
